requirements.txt
# Core LangChain and AI
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
langgraph==0.0.20
langsmith==0.1.0

# Vector Database
pinecone-client==3.0.0
pinecone-text==0.1.0
langchain-pinecone==0.0.3

# Embeddings
sentence-transformers==2.3.1
openai==1.12.0

# API and Web Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Utilities
numpy==1.26.3
pandas==2.1.4
tiktoken==0.5.2

# Evaluation
langchain-evaluation==0.0.8
rouge-score==0.1.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Logging and Monitoring
loguru==0.7.2
